Added `warning` and `catcher` functions.
`warning` is identical to `assert` except that it sends the error to the console instead of throwing an exception.

`catcher` takes a function and invokes it immediately. If the function throws or returns a promise that rejects, the error will be sent to the console. The function result is passed through.

Both functions will only report to console if NODE_ENV is not “production”.